// examples/next-app-dir/src/sdk-src/components/render-content/render-content.types.ts
import type {
  BuilderRenderContext,
  RegisteredComponent,
  BuilderRenderState,
} from '../../context/types'
import type { BuilderContent } from '../../types/builder-content'
import type { Nullable } from '../../types/typescript'
import type { ApiVersion } from '../../types/api-version'

export interface RenderContentProps {
  content?: Nullable<BuilderContent>
  model?: string
  data?: { [key: string]: any }
  context?: BuilderRenderContext
  apiKey: string
  apiVersion?: ApiVersion
  customComponents?: RegisteredComponent[]
  canTrack?: boolean
  locale?: string
  /** @deprecated use `enrich` instead **/
  includeRefs?: boolean
  enrich?: boolean
  /**
   * TO-DO: improve qwik generator to not remap this name for non-HTML tags, then name it `className`
   */
  classNameProp?: string
  hideContent?: boolean
  parentContentId?: string
  isSsrAbTest?: boolean
}

export interface BuilderComponentStateChange {
  state: BuilderRenderState
  ref: {
    name?: string
    props?: {
      builderBlock?: {
        id?: string
      }
    }
  }
}